Split out C-structure wrappers for Python from utils.py into struct.py
